+++ cloud/objstore/objcli/errors.go
@@ -1,6 +1,9 @@
 package objcli
 
-import "errors"
+import (
+	"errors"
+	"fmt"
+)
 
 var (
 	// ErrIncludeAndExcludeAreMutuallyExclusive is returned if the user attempts to supply both the include and exclude
@@ -11,3 +14,11 @@ var (
 	// require upload ids.
 	ErrExpectedNoUploadID = errors.New("received an unexpected upload id, cloud provider doesn't required upload ids")
 )
+
+type ErrVersionUnexpectedType struct {
+	TypeName string
+}
+
+func (e ErrVersionUnexpectedType) Error() string {
+	return fmt.Sprintf("the version field is of an unexpected type, expected %s", e.TypeName)
+}