+++ cloud/objstore/objcli/objaws/mock_service_api.go
@@ -273,6 +273,43 @@ func (_m *mockServiceAPI) HeadObject(ctx context.Context, params *s3.HeadObjectI
 	return r0, r1
 }
 
+// ListObjectVersions provides a mock function with given fields: _a0, _a1, _a2
+func (_m *mockServiceAPI) ListObjectVersions(_a0 context.Context, _a1 *s3.ListObjectVersionsInput, _a2 ...func(*s3.Options)) (*s3.ListObjectVersionsOutput, error) {
+	_va := make([]interface{}, len(_a2))
+	for _i := range _a2 {
+		_va[_i] = _a2[_i]
+	}
+	var _ca []interface{}
+	_ca = append(_ca, _a0, _a1)
+	_ca = append(_ca, _va...)
+	ret := _m.Called(_ca...)
+
+	if len(ret) == 0 {
+		panic("no return value specified for ListObjectVersions")
+	}
+
+	var r0 *s3.ListObjectVersionsOutput
+	var r1 error
+	if rf, ok := ret.Get(0).(func(context.Context, *s3.ListObjectVersionsInput, ...func(*s3.Options)) (*s3.ListObjectVersionsOutput, error)); ok {
+		return rf(_a0, _a1, _a2...)
+	}
+	if rf, ok := ret.Get(0).(func(context.Context, *s3.ListObjectVersionsInput, ...func(*s3.Options)) *s3.ListObjectVersionsOutput); ok {
+		r0 = rf(_a0, _a1, _a2...)
+	} else {
+		if ret.Get(0) != nil {
+			r0 = ret.Get(0).(*s3.ListObjectVersionsOutput)
+		}
+	}
+
+	if rf, ok := ret.Get(1).(func(context.Context, *s3.ListObjectVersionsInput, ...func(*s3.Options)) error); ok {
+		r1 = rf(_a0, _a1, _a2...)
+	} else {
+		r1 = ret.Error(1)
+	}
+
+	return r0, r1
+}
+
 // ListObjectsV2 provides a mock function with given fields: _a0, _a1, _a2
 func (_m *mockServiceAPI) ListObjectsV2(_a0 context.Context, _a1 *s3.ListObjectsV2Input, _a2 ...func(*s3.Options)) (*s3.ListObjectsV2Output, error) {
 	_va := make([]interface{}, len(_a2))